requirements.txt
fastapi==0.118.3
uvicorn==0.38.0
rich==14.2.0
websockets==15.0.1
supabase==2.24.0
python-dotenv==1.2.1
strands-agents[gemini]